+++ lib/flutter_qjs.dart
@@ -5,7 +5,7 @@ import 'dart:isolate';
 import 'dart:typed_data';
 import 'package:ffi/ffi.dart';
 import 'src/ffi.dart';
-export 'src/ffi.dart' show JSEvalFlag;
+export 'src/ffi.dart' show JSEvalFlag, JSRef;
 
 part 'src/engine.dart';
 part 'src/isolate.dart';